requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
config==0.5.0.post0
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
docopt==0.6.2
idna==2.10
pipreqs==0.4.10
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
yarg==0.1.9
